test: Separei createQuestions dos demais testes
