+++ src/main/java/dev/odroca/api_provas/dto/questions/CreateQuestionsResponseDTO.java
@@ -5,12 +5,8 @@ import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-@AllArgsConstructor
-@Getter
-public class CreateQuestionsResponseDTO {
-
-    private UUID id;
-    private int totalCreatedQuestions;
-    private String message;
-    
-}
+public record CreateQuestionsResponseDTO(
+    UUID id,
+    int totalCreatedQuestions,
+    String message
+) {}
