+++ src/test/java/dev/odroca/api_provas/service/QuestionServiceTest.java
@@ -54,7 +54,6 @@ import dev.odroca.api_provas.repository.TestRepository;
 @ExtendWith(MockitoExtension.class)
 public class QuestionServiceTest {
 
-    @Spy
     @InjectMocks
     QuestionService questionService;
 
@@ -63,9 +62,6 @@ public class QuestionServiceTest {
 
     @Mock
     QuestionRepository questionRepository;
-
-    @Mock
-    OptionRepository optionRepository;
 
     @Mock
     QuestionMapper questionMapper;
@@ -206,39 +202,6 @@ public class QuestionServiceTest {
         verify(testRepository).findById(testId);
         verify(optionMapper).createDtoToEntityList(options);
         verifyNoInteractions(questionRepository);
-    }
-
-    @Test
-    @DisplayName("Deve criar duas questões quando tudo estiver OK")
-    void createQuestionsSuccessful() {
-        int totalQuestions = 2;
-        UUID testId = UUID.fromString("5e6863bc-4f69-4a95-b672-c41296ec95a2");
-
-        Set<CreateOptionModelDTO> options = new HashSet<>();
-        for (Integer i = 0; i < 5; i++) {
-            Boolean isCorrect = (i == 4);
-            options.add(new CreateOptionModelDTO(i.toString(), isCorrect));
-        }
-
-        Set<CreateQuestionModelDTO> questions = new HashSet<>();
-        questions.add(new CreateQuestionModelDTO("1+1?", options));
-        questions.add(new CreateQuestionModelDTO("6-4?", options));
-        
-        CreateQuestionsRequestDTO questionsModel = new CreateQuestionsRequestDTO(questions);
-
-        // Não faço nada com o response de createQuestion
-        doReturn(null).when(questionService).createQuestion(any(UUID.class), any(CreateQuestionModelDTO.class));
-        CreateQuestionsResponseDTO result = questionService.createQuestions(testId, questionsModel);
-
-        assertNotNull(result);  
-        assertNotNull(result.getId());
-        assertNotNull(result.getTotalCreatedQuestions());
-        assertEquals(testId, result.getId());
-        assertEquals(totalQuestions, result.getTotalCreatedQuestions());
-
-        // Verifica se realmente o looping de createQuestions chama createQuestion 2x
-        verify(questionService, times(2)).createQuestion(eq(testId), any(CreateQuestionModelDTO.class));
-        
     }
 
     @Test
